Demo/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.0.4
heroku==0.1.4n
idna==2.8
numpy==1.17.4
opencv-python==4.2.0.32
psycopg2==2.8.4
python-dateutil==1.5
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3====1.21.1
whitenoise==5.0.1
